backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Task queue
celery[redis]==5.4.0
redis==5.2.0
flower==2.0.1

# Image processing
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.1.3

# Auth
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Email
aiosmtplib==3.0.1

# Logging
structlog==24.4.0

# HTTP + file handling
httpx==0.27.2
python-multipart==0.0.12
aiofiles==24.1.0

# Validation + settings
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.1.1
